added create and get orders APIs

+++ database/migrations/20240316072317-create-order_items-table.js
@@ -20,7 +20,7 @@ module.exports = {
         allowNull: false,
       },
       amount: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(10, 2),
         allowNull: false,
       },
       archived: {
